feat!(misc): add `misc` module
BREAKING CHANGE
to handle custom traits as `IsClose` or
appo enum as `ByteOrder`

refactor all other modules
