+++ src/main.rs
@@ -2,6 +2,7 @@ mod cli;
 mod color;
 mod error;
 mod hdrimage;
+mod misc;
 mod normal;
 mod point;
 mod transformation;
